Add files via upload
The improved version of Assignment2. It has combined the modes to fit in a new mode, survival  :)
Similarly allows user to pick a mode, and then accepts inputs for randomly generated math questions
Has the option to play either addition & subtraction, multiplication & division, or survival mode which uses all 4. 
The operation given to the user is randomly generated. For division, a special generation was used, unique from the others in order to prevent awkward decimal numbers. It will loop until it creates an equation in which the result is a whole number. Some of the multiplication ones are admittedly pretty hard due to the high numbers  :P